// src/shared/api/models/getApiServicesDeposit200DataGroupedServicesItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiServicesDeposit200DataGroupedServicesItemServicesItem } from './getApiServicesDeposit200DataGroupedServicesItemServicesItem';

export type GetApiServicesDeposit200DataGroupedServicesItem = {
    /** Group ID (0 for ungrouped services) */
    id?: number;
    /** Group name ("Ungrouped services" for services without group) */
    group?: string;
    /**
     * Group creation date
     * @nullable
     */
    createdAt?: string | null;
    /**
     * Group sequence order
     * @nullable
     */
    sequence?: number | null;
    services?: GetApiServicesDeposit200DataGroupedServicesItemServicesItem[];
};